Putting through the first pass of an autocomplete command.
It's starting to work well, already realised that the default behaviour
should not be to merge all of the autocompleted files together.

You should probably autocomplete on a per-dictionary basis and then add
those to plover individually with individual priority (probably the
entry right below the actual dictionary.

For example, in lapwing-base vs lapwing-uk-additions, some of the
autocorrects for lapwing-uk-additions SHOULD override the entries for
lapwing-base, like UPB/ORGD.